+++ src/components/Navbar/PlanEngage.js
@@ -2,8 +2,8 @@ import React from 'react'
 import cx from 'classnames'
 import { Link } from 'react-router-dom'
 import Icon from '@santiment-network/ui/Icon'
-import { useUser } from '../../contexts/user'
-import { useUserSubscriptionStatus } from '../../contexts/user/subscriptions'
+import { useUser } from '../../stores/user'
+import { useUserSubscriptionStatus } from '../../stores/user/subscriptions'
 import styles from './PlanEngage.module.scss'
 
 export const getTrialDaysLeft = () => ''
