+++ src/ducks/SANCharts/ChartSettingsContextMenu.js
@@ -8,7 +8,7 @@ import Label from '@santiment-network/ui/Label'
 import UIIcon from '@santiment-network/ui/Icon'
 import Panel from '@santiment-network/ui/Panel/Panel'
 import ShareModalTrigger from '../../components/Share/ShareModalTrigger'
-import { useUserSubscriptionStatus } from '../../contexts/user/subscriptions'
+import { useUserSubscriptionStatus } from '../../stores/user/subscriptions'
 import ChartDownloadBtn from './ChartDownloadBtn'
 import DownloadCSVBtn from './DownloadCSVBtn'
 import styles from './ChartPage.module.scss'
