+++ src/pages/Pricing/Plans.js
@@ -4,8 +4,8 @@ import Toggle from '@santiment-network/ui/Toggle'
 import Plan from './Plan'
 import { noBasicPlan } from '../../utils/plans'
 import { usePlans } from '../../ducks/Plans/hooks'
-import { useUser } from '../../contexts/user'
-import { useUserSubscription } from '../../contexts/user/subscriptions'
+import { useUser } from '../../stores/user'
+import { useUserSubscription } from '../../stores/user/subscriptions'
 import Enterprise from './Enterprise'
 import styles from './Plans.module.scss'
