+++ src/stores/ui/index.js
@@ -1,0 +1,1 @@
+export function useIsLoggedIn () {}
